(BSR)[API] feat: move venues endpoints in separate file
